Rename types.hpp - the one in TEC, not VComputer
